warnings resolvidos e release

+++ JogoTecProg/CuboCowboy.cpp
@@ -35,7 +35,7 @@ void CuboCowboy::mover(double t)
         v.y = 0;
     }
     else {
-        v.y += g * t; //Gravidade
+        v.y += g * (float)t; //Gravidade
     }
 
     delay += t;
